pyproject.toml
[tool.poetry]
name = "sv-oidc"
version = "0.1.0"
description = ""
authors = ["JoseVL92 <jovalab92@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.110.0"
pymongo = "4.6.1"
python-jose = "3.3.0"
jwcrypto = "1.5.1"
cuid2 = "2.0.0"
cryptography = "42.0.5"
python-dotenv = "1.0.1"
uvicorn = "0.27.1"
pydantic = {extras = ["email"], version = "^2.10.6"}
redis = {extras = ["hiredis"], version = "^5.2.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
